fix: user not switching when entering data collection epic
